[FEAT]: FeedHistory fetch해오는 메서드 FeedAPI에서 MyPageAPI로 변경
