+++ Projects/Domain/Repository/Interfaces/MyPageRepository.swift
@@ -13,4 +13,5 @@ import Entity
 public protocol MyPageRepository {
   func fetchMyPageSummary() -> Single<MyPageSummary>
   func fetchVerifiedChallengeDates() -> Single<[Date]>
+  func fetchFeedHistory(page: Int, size: Int) async throws -> PaginationResultType<FeedHistory>
 }
